Fanatically correct real tensor cloning for propagate_real_tensors
Internal xref:
https://fb.workplace.com/groups/6829516587176185/posts/7211398545654652/

Previously I did it in a crappy way using clone_input in the callback,
but this results in tensors that don't have quite the same
size/stride/storage offset and there was an internal test case where
not having completely accurate information was causing a downstream
problem in propagation.  So now I make real tensors as similar to their
fake equivalents as much as possible.  Though... I don't bother with
autograd lol.

Signed-off-by: Edward Z. Yang <ezyang@meta.com>

ghstack-source-id: c35e5051ec3c9a5c6b8da9102d866d70ed11274d
Pull Request resolved: https://github.com/pytorch/pytorch/pull/126175