[inductor] Fix ops.scan for non-commutative operators
`tl.associative_scan` supports non-commutative combine functions but `tl.reduce`
doesn't. This effects non-persistent scans, where we use the reduction from
the previous loop iterations as the base for future iterations.

Here I work around this by taking the last element of the scan output and using
that as the reduced value. This is possibly by using a trick where we create a
mask that is 1 at the desired element and 0 elsewhere, then sum over it.

ghstack-source-id: 23a6f1b7717633dcf38743aad9d5b6bc937dc935
Pull Request resolved: https://github.com/pytorch/pytorch/pull/126633