[c10d] fix excepthook crash on exc after destroy_process_group
fixes #126379

This is the easy fix.  An additional fix that I did not do is to
deregister the excepthook (or rather, restore the orignal one) when
calling dist.destroy_process_group.  This might be a bit complicated in
practice, so landing as is for now.

Also, couldn't figure out a clean way to test this.  assertRaisesRegex
wasn't getting a string value, probably becuase of the stderr
redirection done via the excepthook in the first place.

ghstack-source-id: 2ffe93e104eec809881411120f08a37b9285fe16
Pull Request resolved: https://github.com/pytorch/pytorch/pull/126739