[pipelining] Validate stage input/output shape/dtype
Address the classes of user errors stemming from (possibly)
unintentional dynamic shapes usage or mismatch of configuration time and
run time data shapes/dtypes.

The goal is to ensure a clear error is raised rather than relying on some underlying
error to bubble up when a tensor shape is not compatible, or worse,
having a silent correctness issue.

**Classes of shape/dtype errors**
(a) error is thrown within the stage-module forward code, but may be
hard to understand/trace back to an input issue
(b) silent correctness issue happens inside the stage-module forward,
but the correct output shape is still produced
produces the expected output shape
(c) the stage-module produces an output that is locally correct, but not
matching the expectation of the following stage, leading to a hang or
correctness issue down the line

**How validation helps**
Input shape validation
- improves debugability of case (a)
- guards against case (b)
- only needed on first stage, since subsequent stages use pre-allocated recv
  buffers that can't change shape/size even if they wanted to

Output shape validation
- guards against case (c)

Validation of first stage input and all stages' outputs inductively verifies all shapes

Shape/dtype are most critical as they literally affect the number of
bytes on the wire.  Strides and other tensor properties may also (?)
matter, and the validation function can be adjusted accordingly if needed.

ghstack-source-id: 6f7c523747af13ce82bda6ed47a01eb70335da96
Pull Request resolved: https://github.com/pytorch/pytorch/pull/126732