[AOTI] Update reinplace to cover mutated buffer
Summary: Unlike JIT Inductor, AOTI currently unlifts weights and buffers from input args, so the reinplace pass didn't really work for AOTI because it only checks mutation on placeholder, which led to excessive memory copies for kv_cache updates in LLM models. This PR removes those memory copies and roughly offers a 2x speedup. In the future, we will revert the unlift logic in AOTI and make the behvior consitent with JIT Inductor.

ghstack-source-id: 8cd91b30855169425d75207d03e236c98ce03c0e
Pull Request resolved: https://github.com/pytorch/pytorch/pull/127295