Save quantization_tag in export graph serialization (#127473)
Summary:

`quantization_tag` is a first class citizen metadata in quantization flows that is preserved by it. As we'll want to store the quantized exported graphs we also need to preserve this metadata as it's used in later flows. Only json supported metadata will be allowed to be serialized.

Test Plan: Added test case

Reviewed By: angelayi

Differential Revision: D57939282